Adjusted the nuke-adapter code so that it can be shared more easily with Houdini
