+++ _environment.py
@@ -13,6 +13,7 @@ rezzurect will fail to build / source / run Rez packages correctly.
 import platform
 
 # IMPORT LOCAL LIBRARIES
+from .adapters.houdini_installation import houdini_installation_builder
 from .adapters.nuke_installation import nuke_installation_builder
 from .utils import common
 
@@ -48,6 +49,7 @@ def main(
 
     '''
     known_modules = (
+        houdini_installation_builder,
         nuke_installation_builder,
     )
