+++ adapters/nuke_installation/nuke_installation_builder.py
@@ -13,7 +13,6 @@ import stat
 import os
 
 # IMPORT LOCAL LIBRARIES
-from ...strategies import internet
 from ...utils import progressbar
 from .. import base_builder
 from ... import chooser
@@ -21,7 +20,7 @@ from . import helper
 
 
 _DEFAULT_VALUE = object()
-_LOGGER = logging.getLogger('rezzurect.nuke_builder')
+_LOGGER = logging.getLogger('rezzurect.nuke_installation_builder')
 
 
 class BaseNukeAdapter(base_builder.BaseAdapter):
@@ -311,94 +310,6 @@ class WindowsAdapter(BaseNukeAdapter):
                                ''.format(stderr=stderr))
 
 
-def add_local_filesystem_build(source_path, install_path, adapter):
-    '''Search the user's files and build the Rez package.
-
-    Args:
-        adapter (`rezzurect.adapters.base_builder.BaseAdapter`):
-            The object which is used to "install" the files.
-        source_path (str):
-            The absolute path to where the Rez package is located, on-disk.
-        install_path (str):
-            The absolute path to where the package will be installed into.
-
-    '''
-    if not os.path.isdir(install_path):
-        os.makedirs(install_path)
-
-    adapter.install_from_local(source_path, install_path)
-
-
-def add_link_build(adapter):
-    '''Add the command which lets the user link Rez to an existing install.
-
-    Args:
-        adapter (`rezzurect.adapters.base_builder.BaseAdapter`):
-            The object which is used to search for existing installs.
-
-    Raises:
-        RuntimeError: If no valid executable could be found.
-
-    '''
-    paths = adapter.get_preinstalled_executables()
-
-    for path in paths:
-        if os.path.isfile(path):
-            return
-
-    raise RuntimeError('No expected binary file could be found. '
-                       'Checked "{paths}".'.format(paths=', '.join(sorted(paths))))
-
-
-def add_from_internet_build(package, system, distribution, architecture, source_path, install_path, adapter):
-    '''Download the installer for `package` and then install it.
-
-    Args:
-        package (str):
-            he name of packaget to get an installer from online.
-        system (str):
-            The name of the OS platform. Example: "Linux", "Windows", etc.
-        distribution (str):
-            The name of the type of OS (Example: "CentOS", "windows", etc.)
-        architecture (str):
-            The bits of the `system`. Example: "x86_64", "AMD64", etc.
-        source_path (str):
-            The absolute path to where the Rez package is located, on-disk.
-        install_path (str):
-            The absolute path to where the package will be installed into.
-        adapter (`rezzurect.adapters.base_builder.BaseAdapter`):
-            The object which is used to "install" the files.
-
-    Raises:
-        RuntimeError: If the download failed to install into `destination`.
-
-    '''
-    destination = adapter.get_archive_folder(source_path)
-
-    destination = internet.download(
-        package,
-        adapter.version,
-        system,
-        distribution,
-        architecture,
-        destination,
-    )
-
-    if not os.path.isfile(destination):
-        raise RuntimeError(
-            'Package/Version "{package}/{adapter.version}" could not be downloaded to path, '
-            '"{destination}".'.format(package=package, adapter=adapter, destination=destination))
-
-    _LOGGER.info(
-        'Downloaded package/version "%s/%s" to path, "%s".',
-        package,
-        adapter.version,
-        destination,
-    )
-
-    add_local_filesystem_build(source_path, install_path, adapter)
-
-
 def register(source_path, install_path, system, distribution, architecture):
     '''Add installation options to all of the Nuke adapter classes.
 
@@ -429,13 +340,13 @@ def register(source_path, install_path, system, distribution, architecture):
         adapter.strategies = []
 
         add_nuke_from_internet_build = functools.partial(
-            add_from_internet_build,
+            base_builder.add_from_internet_build,
             'nuke', system, distribution, architecture, source_path, install_path)
         add_nuke_local_filesystem_build = functools.partial(
-            add_local_filesystem_build, source_path, install_path)
+            base_builder.add_local_filesystem_build, source_path, install_path)
 
         adapter.strategies.append(('local', add_nuke_local_filesystem_build))
-        adapter.strategies.append(('link', add_link_build))
+        adapter.strategies.append(('link', base_builder.add_link_build))
         adapter.strategies.append(('internet', add_nuke_from_internet_build))
 
         chooser.register_build_adapter(adapter, 'nuke_installation', system_)
